+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - Added a "Component" element shape.
 - Added a "Dotted" element border style.
 - Components from any container can now be added to a component view.
+- Added the ability to customize the symbols used when rendering metadata.
 
 ## 0.9.6 (29th February 2020)